@@ -3,118 +3,135 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@babel/polyfill": "^7.4.3",
     "@celebryts/react-autocomplete-tags": "^1.0.0",
-    "@fortawesome/fontawesome": "^1.1.4",
-    "@fortawesome/fontawesome-free": "^5.3.1",
-    "@fortawesome/fontawesome-free-solid": "^5.0.8",
-    "@fortawesome/react-fontawesome": "0.0.17",
+    "@fortawesome/fontawesome": "^1.1.8",
+    "@fortawesome/fontawesome-free": "^5.8.1",
+    "@fortawesome/fontawesome-free-solid": "^5.0.13",
+    "@fortawesome/fontawesome-svg-core": "^1.2.17",
+    "@fortawesome/react-fontawesome": "0.1.4",
     "@types/react-dnd": "^3.0.2",
-    "avsc": "^5.0.4",
-    "babel-preset-es2015": "^6.24.1",
+    "avsc": "^5.4.7",
     "base-64": "^0.1.0",
-    "bootstrap": "^4.1.2",
+    "bootstrap": "^4.3.1",
     "brace": "^0.11.1",
-    "chart.js": "^1.1.1",
-    "clean-webpack-plugin": "^0.1.19",
-    "copy-webpack-plugin": "^4.5.2",
+    "chart.js": "^2.8.0",
+    "clean-webpack-plugin": "^2.0.1",
+    "copy-webpack-plugin": "^5.0.2",
+    "css-loader": "^2.1.1",
     "downloadjs": "^1.4.7",
-    "draft-js": "^0.10.1",
-    "draftjs-to-html": "^0.7.5",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "flag-icon-css": "^3.2.1",
-    "fusioncharts": "^3.13.1-sr.2",
+    "draft-js": "^0.10.5",
+    "draftjs-to-html": "^0.8.4",
+    "file-loader": "^3.0.1",
+    "flag-icon-css": "^3.3.0",
     "html-loader": "^0.5.5",
-    "html-to-draftjs": "0.1.0-beta14",
-    "immutable": "^3.8.1",
-    "jquery": "^3.2.1",
+    "html-to-draftjs": "1.4.0",
+    "html-webpack-plugin": "^3.2.0",
+    "immutable": "^3.8.2",
+    "jquery": "^3.3.1",
     "jwt-decode": "^2.2.0",
-    "localforage": "^1.7.2",
-    "moment": "^2.21.0",
-    "papaparse": "^4.3.3",
-    "prop-types": "^15.5.6",
-    "rc-collapse": "^1.9.3",
+    "localforage": "^1.7.3",
+    "moment": "^2.24.0",
+    "papaparse": "^4.6.3",
+    "query-string": "^6.4.2",
+    "rc-collapse": "^1.11.1",
     "rc-steps": "^3.3.1",
-    "react": "^15.6.1",
-    "react-ace": "^6.2.0",
-    "react-autocomplete": "^1.7.2",
-    "react-autosuggest": "^9.4.2",
-    "react-chartjs": "^0.8.0",
+    "react": "^16.8.6",
+    "react-ace": "^6.4.0",
+    "react-autocomplete": "^1.8.1",
+    "react-autosuggest": "^9.4.3",
+    "react-chartjs": "^1.2.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-datepicker": "^1.2.2",
-    "react-dates": "^16.3.6",
+    "react-datepicker": "^2.3.0",
+    "react-dates": "^20.1.0",
     "react-dazzle": "^1.4.0",
-    "react-dimension": "^0.1.1",
-    "react-dimensions": "^1.3.0",
-    "react-dnd": "^2.6.0",
-    "react-dnd-html5-backend": "^2.4.1",
-    "react-dom": "^15.6.1",
+    "react-dnd": "^7.4.4",
+    "react-dnd-html5-backend": "^7.4.4",
+    "react-dom": "^16.8.6",
     "react-draft": "0.0.0",
     "react-draft-js": "^1.0.0",
-    "react-draft-wysiwyg": "^1.10.8",
+    "react-draft-wysiwyg": "^1.13.2",
     "react-drag-drop-container": "^6.1.1",
-    "react-dropzone": "^3.13.4",
-    "react-dropzone-component": "^2.0.0",
-    "react-event-timeline": "^1.3.3",
+    "react-dropzone": "^10.1.0",
+    "react-dropzone-component": "^3.2.0",
+    "react-event-timeline": "^1.6.3",
     "react-grid-layout": "^0.16.6",
     "react-infinite-scroll": "^0.1.5",
-    "react-json-view": "^1.15.2",
+    "react-json-view": "^1.19.1",
     "react-medium-editor": "^1.8.1",
-    "react-modal": "^2.2.2",
+    "react-modal": "^3.8.1",
     "react-modal-bootstrap": "^1.1.1",
     "react-overlay-loading": "^1.0.3",
     "react-querybuilder": "^1.4.3",
     "react-recaptcha": "^2.3.10",
-    "react-redux": "^5.0.5",
-    "react-redux-toastr": "^7.2.3",
+    "react-redux": "^6.0.1",
+    "react-redux-toastr": "^7.4.8",
     "react-resizable": "^1.7.5",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "react-select": "^1.0.0-rc.5",
-    "react-sortable-tree": "^2.2.0",
-    "react-stepzilla": "^4.7.2",
-    "react-syntax-highlighter": "^10.1.3",
-    "react-table": "^6.8.6",
-    "react-tag-autocomplete": "^5.6.0",
-    "react-tag-input": "^4.7.2",
-    "react-through": "^1.1.0",
-    "react-timestamp": "^4.3.1",
-    "react-transition-group": "^2.2.1",
-    "reactstrap": "^6.2.0",
-    "recharts": "^1.4.1",
-    "redux": "^3.7.0",
-    "redux-file-upload": "0.0.18",
-    "redux-form": "^6.8.0",
+    "react-router": "^5.0.0",
+    "react-router-dom": "^5.0.0",
+    "react-select": "^2.4.2",
+    "react-sortable-tree": "^2.6.2",
+    "react-stepzilla": "^6.0.0",
+    "react-syntax-highlighter": "^10.2.1",
+    "react-table": "^6.9.2",
+    "react-tag-autocomplete": "^5.9.0",
+    "react-tag-input": "^6.4.0",
+    "react-through": "^1.1.3",
+    "react-timestamp": "^5.0.0",
+    "react-transition-group": "^2.8.0",
+    "reactstrap": "^7.1.0",
+    "recharts": "^1.5.0",
+    "redux": "^4.0.1",
+    "redux-file-upload": "0.0.19",
+    "redux-form": "^8.1.0",
     "redux-form-dropzone": "0.0.1",
-    "redux-form-website-template": "0.0.123",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
-    "serviceworker-webpack-plugin": "^0.2.3",
+    "redux-thunk": "^2.3.0",
+    "serviceworker-webpack-plugin": "^1.0.1",
     "set-default-browser": "^1.0.1",
-    "shortid": "^2.2.13",
-    "socket.io-client": "^2.1.1",
-    "superagent": "^3.8.3",
-    "superagent-mock": "^3.4.0",
+    "shortid": "^2.2.14",
+    "socket.io-client": "^2.2.0",
+    "style-loader": "^0.23.1",
+    "superagent": "^5.0.2",
+    "superagent-mock": "^3.7.0",
     "superagent-promise": "^1.1.0",
-    "throttle-debounce": "^1.0.1",
-    "webpack": "^3.11.0",
+    "throttle-debounce": "^2.1.0",
     "xml2js": "^0.4.19"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.2.1",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-es2015-template-literals": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "concurrently": "^3.5.0",
-    "mock-json-server": "0.0.4",
-    "react-scripts": "^1.1.4",
-    "uglifyjs-webpack-plugin": "^1.2.7",
-    "webpack": "^3.10.0",
-    "webpack-merge": "^4.1.1"
+    "@babel/core": "^7.4.3",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-do-expressions": "^7.0.0",
+    "@babel/plugin-proposal-export-default-from": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-function-bind": "^7.0.0",
+    "@babel/plugin-proposal-function-sent": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.0.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/plugin-transform-template-literals": "^7.0.0",
+    "@babel/preset-env": "^7.4.3",
+    "@babel/preset-react": "^7.0.0",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.5",
+    "concurrently": "^4.1.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "mock-json-server": "1.3.2",
+    "path": "^0.12.7",
+    "uglifyjs-webpack-plugin": "^2.1.2",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.0",
+    "webpack-dev-server": "^3.2.1",
+    "webpack-merge": "^4.2.1"
   },
   "scripts": {
     "start": "webpack-dev-server --progress",
