@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { connect } from 'react-redux'
 import { reset } from 'redux-form'
 import fontawesome from '@fortawesome/fontawesome'
-import FontAwesomeIcon from '@fortawesome/react-fontawesome'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faGlobe } from '@fortawesome/fontawesome-free-solid'
 import {
 	loadDatasets,
@@ -43,7 +43,7 @@ class Header extends Component {
       this.pushToPublic = this.pushToPublic.bind(this)
 		}
   
-    componentWillReceiveProps(nextProps){
+    UNSAFE_componentWillReceiveProps(nextProps){
       
       if(nextProps.notifications){
         var unreadNot = nextProps.notifications.filter(notification =>{
@@ -219,6 +219,16 @@ class Header extends Component {
 			let crea = this.state.crea ? "show" : ""
 			var navigation = this.state.navigation?" active":""
 			var show = this.state.navigation?" show":"" 
+
+			var iconStyle = {
+				display: 'inline-block',
+				width: '20px',
+				marginRight: '10px',
+				marginLeft: '-10px',
+				color: '#c2cfd6',
+				textAlign: 'center',
+			}
+
 			return (
 				<header className="app-header navbar border-0">
 				<button className="nav-link navbar-toggler sidebar-toggler d-lg-none" onClick={this.mobileSidebarToggle} type="button">&#9776;</button>
@@ -247,10 +257,10 @@ class Header extends Component {
 								<button className="w-100 h-100 btn btn-header" onClick={/* this.crea.bind(this) */this.toggleCrea}><i className="fa fa-plus fa-lg"/></button>
 								<div className={"dropdown-menu m-0 dropdown-menu-right "+ crea} aria-labelledby="dropdownMenuButton">
 									<h6 className="dropdown-header text-center"><b>Crea</b></h6>
-									{(isEditor(loggedUser) || isAdmin(loggedUser)) && <button className="dropdown-item" onClick={()=> { const{ dispatch } = this.props; dispatch(reset('wizard')); this.props.history.push('/private/ingestionwizzard'); this.toggleCrea}}><i className="fa fa-table"></i> Nuovo Dataset</button>}
+									{(isEditor(loggedUser) || isAdmin(loggedUser)) && <button className="dropdown-item" onClick={()=> { const{ dispatch } = this.props; dispatch(reset('wizard')); this.props.history.push('/private/ingestionwizzard'); this.toggleCrea}}><i className="fa fa-table" style={iconStyle}></i> Nuovo Dataset</button>}
 									{/* <button className="dropdown-item" onClick={this.createWidget} ><i className="fa fa-chart-bar"></i> Nuovo Widget</button> */}
-									<button className="dropdown-item" onClick={this.createDash} ><i className="fa fa-columns"></i> Nuova Dashboard</button>
-									<button className="dropdown-item" onClick={this.createStory} ><i className="fa fa-font"></i> Nuova Storia</button>
+									<button className="dropdown-item" onClick={this.createDash} ><i className="fa fa-columns" style={iconStyle}></i> Nuova Dashboard</button>
+									<button className="dropdown-item" onClick={this.createStory} ><i className="fa fa-font" style={iconStyle}></i> Nuova Storia</button>
 							</div>
 						</div>
 						{/* <button className="w-100 h-100 btn btn-header" onClick={this.crea.bind(this)}><i className="fa fa-plus fa-lg"/></button> */}
@@ -271,8 +281,8 @@ class Header extends Component {
 								</a>
 								<div className={"dropdown-menu dropdown-menu-right "+ open} aria-labelledby="dropdownMenuButton">
 									<h6 className="dropdown-header text-center"><b>{loggedUser ? loggedUser.givenname : ''}</b></h6>
-									<a className="dropdown-item" href="/#/private/profile" onClick={this.toggle}><i className="fa fa-user"></i> Profilo</a>
-									<a className="dropdown-item" onClick={() => { logout(); this.toggle }} href="/"><i className="fa fa-lock"></i> Logout</a>
+									<a className="dropdown-item" href="/#/private/profile" onClick={this.toggle}><i className="fa fa-user" style={iconStyle}></i> Profilo</a>
+									<a className="dropdown-item" onClick={() => { logout(); this.toggle }} href="/"><i className="fa fa-lock" style={iconStyle}></i> Logout</a>
 							</div>
 						</div>
 					</li>
@@ -283,7 +293,7 @@ class Header extends Component {
 						<button className={"h-100 btn btn-accento"+navigation} id="dropdown" data-toggle="dropdown"  aria-haspopup="true" aria-expanded="false" onClick={this.toggleNav.bind(this)}><p className="m-0 p-0 d-md-down-none float-left">Area Privata</p><i className="float-left fa fa-lock fa-lg d-lg-none"/> <i className="fa fa-sort-down ml-2 align-top"/></button>
 							<div className={"dropdown-menu dropdown-menu-right m-0" + show} aria-labelledby="dropdownMenuButton">
 								<h6 className="dropdown-header bg-white"><b>VAI A</b></h6>
-								<button className="dropdown-item bg-light b-l-pvt border-primary pr-5" onClick={this.pushToPublic}>
+								<button className="dropdown-item bg-light b-l-pvt border-primary pr-5" onClick={this.pushToPublic.bind(this)}>
 										<div className="row">
 												<h5 className="col-1 pl-0"><FontAwesomeIcon icon={faGlobe} className="mx-2"/></h5>
 												<div className="row col-11 ml-1">
