@@ -1,17 +1,9 @@
 import React from 'react';
-import {
-  Modal,
-  ModalHeader,
-  ModalTitle,
-  ModalClose,
-  ModalBody,
-  ModalFooter
-} from 'react-modal-bootstrap'
 import { isEditor, isAdmin } from '../../../../utility'
 import { connect } from 'react-redux'
 import { Tooltip } from 'reactstrap';
 import fontawesome from '@fortawesome/fontawesome'
-import FontAwesomeIcon from '@fortawesome/react-fontawesome'
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import { faGlobe, faUsers, faSortDown, faUser } from '@fortawesome/fontawesome-free-solid'
 
 class EditBarTop extends React.Component {
@@ -33,7 +25,7 @@ class EditBarTop extends React.Component {
     this.condividi = this.condividi.bind(this);
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if(nextProps.dashboard)
       this.state.dashboard = nextProps.dashboard;
     if(nextProps.modified)
