@@ -17,7 +17,7 @@ class TextEditor extends Component {
     this.handleChange = this.handleChange.bind(this);
   }
   
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     this.setState({
       text : nextProps.text
     });
