@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import fontawesome from '@fortawesome/fontawesome'
-import FontAwesomeIcon from '@fortawesome/react-fontawesome'
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import { faLock, faGlobe } from '@fortawesome/fontawesome-free-solid'
 import { truncateDatasetName, decodeTheme, isPublic } from '../../utility' 
 
@@ -60,13 +60,13 @@ class DatasetCard extends Component{
                                 !openData &&
                                 //<span className="badge badge-pill badge-danger fa-pull-right badge-dash my-1" title="Il dataset è privato"> </span>
                                 //<i className="fa fa-lock fa-pull-right fa-lg text-icon my-1 pointer" title='Il dataset è privato'/>
-                                <span className="pointer" title='Il dataset è privato'><i icon={faLock} className="fas fa-lock fa-pull-right text-icon pointer" style={{fontSize: '16px'}}/></span>
+                                <span className="pointer" title='Il dataset è privato'><i className="fas fa-lock fa-pull-right text-icon pointer" style={{fontSize: '16px'}}/></span>
                             }
                             {
                                 openData &&
                                 //<span className="badge badge-pill badge-success fa-pull-right badge-dash my-1" title="Il dataset è pubblico"> </span>
                                 //<i className="fa fa-globe fa-pull-right fa-lg text-icon my-1 pointer" title='Il dataset è pubblico'/>
-                                <span className="pointer" title='Il dataset è pubblico'><i icon={faGlobe} className="fas fa-globe fa-pull-right text-icon pointer" style={{fontSize: '16px'}}/></span>
+                                <span className="pointer" title='Il dataset è pubblico'><i className="fas fa-globe fa-pull-right text-icon pointer" style={{fontSize: '16px'}}/></span>
                             }
                         </div>
                     </div>
