@@ -22,15 +22,11 @@ import DashboardService from './services/DashboardService';
 
 import { serviceurl } from "../../../config/serviceurl";
 
-// We are using bootstrap as the UI library
-// Removed for conflicts
-//import 'bootstrap/dist/css/bootstrap.css';
-
 // Default styes of dazzle.
-import 'react-dazzle/lib/style/style.css';
+require('react-dazzle/lib/style/style.css');
 
 // Our styles
-import '../styles/custom.css';
+require('../styles/custom.css');
 
 
 const widgetService = new WidgetService();
