@@ -5,7 +5,7 @@ import { toastr } from 'react-redux-toastr'
 import { isEditor, isAdmin } from '../../../../utility'
 import { Tooltip } from 'reactstrap';
 import fontawesome from '@fortawesome/fontawesome'
-import FontAwesomeIcon from '@fortawesome/react-fontawesome'
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import { faGlobe, faUsers, faSortDown, faUser } from '@fortawesome/fontawesome-free-solid'
 import DashboardService from '../services/DashboardService';
 
