@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import 'rc-steps/assets/index.css';
-import 'rc-steps/assets/iconfont.css';
+require('rc-steps/assets/index.css');
+require('rc-steps/assets/iconfont.css');
 import Steps, { Step } from 'rc-steps';
 import { messages } from '../../../i18n/i18n-ita';
 
