@@ -5,7 +5,7 @@ import UserStoryView from './components/UserStoryView.jsx';
 import { Switch, Route } from 'react-router-dom';
 
 // Our styles
-import './styles/custom.css';
+require ('./styles/custom.css');
 
 class UserStory extends Component {
   constructor(props) {
