@@ -12,7 +12,7 @@ import Container from './Container.jsx';
 import UserStoryService from './services/UserStoryService';
 
 // Our styles
-import '../styles/custom.css';
+require('../styles/custom.css');
 
 
 const userStoryService = new UserStoryService();
